refactor: QueryLogController 및 DTO 클래스 패키지 경로 변경
1. QueryLogController를 presentation에서 adapter.in.web으로 이동한다.
2. QueryRequest와 QueryResponse를 presentation.dtos에서 application.commandquery로 이동한다.
3. 관련 import 경로를 최신 상태로 수정한다.
4. QueryResponse에 Users 정보를 사용하는 새로운 of 메서드를 추가한다.

Resolves #22
